change double to long double, update main

+++ helper.cpp
@@ -19,7 +19,7 @@ bool point::operator<(const point &other) const
 
 size_t point_hash::operator()(const point &p) const
 {
-    return std::hash<double>()(p.x) ^ std::hash<double>()(p.y);
+    return std::hash<long double>()(p.x) ^ std::hash<long double>()(p.y);
 }
 
 int cmp(const point &a, const point &b)
@@ -34,27 +34,27 @@ int cmp_2(const point &a, const point &b)
     return a.x == b.x ? a.y > b.y : a.x < b.x;
 }
 
-double dis(const point &a, const point &b)
+long double dis(const point &a, const point &b)
 {
     return sqrt((a.x - b.x) * (a.x - b.x) + (a.y - b.y) * (a.y - b.y));
 }
 
-double turn_test(const point &p1, const point &p2, const point &p3)
+long double turn_test(const point &p1, const point &p2, const point &p3)
 // return: left (counter-clockwise): > 0
 {
     return p1.x * (p2.y - p3.y) + p2.x * (p3.y - p1.y) + p3.x * (p1.y - p2.y);
 }
 
-double clockwise_angle(const point &a, const point &b, const point &c)
+long double clockwise_angle(const point &a, const point &b, const point &c)
 // angle between AB and BC in clockwise direction
 {
     //  atan2  return arctan y/x in [−pi , pi]
-    double a_to_b = atan2(b.y - a.y, b.x - a.x);
-    double a_to_c = atan2(c.y - a.y, c.x - a.x);
-    double angle_acb = a_to_b - a_to_c;
+    long double a_to_b = atan2(b.y - a.y, b.x - a.x);
+    long double a_to_c = atan2(c.y - a.y, c.x - a.x);
+    long double angle_acb = a_to_b - a_to_c;
 
     // range in [-pi, pi]
-    const double Pi = acos(-1); // or use some π constant
+    const long double Pi = acos(-1); // or use some π constant
     if (angle_acb > Pi)
         angle_acb -= 2 * Pi;
     else if (angle_acb < -Pi)
